(#7) refactor: structure data for pages
