+++ src/redux/store/store.ts
@@ -2,8 +2,8 @@ import { applyMiddleware, createStore, Store } from 'redux';
 import { composeWithDevTools } from 'redux-devtools-extension';
 import logger from 'redux-logger';
 import createSagaMiddleware from 'redux-saga';
-import { rootReducers } from './reducers';
-import { rootSaga } from './sagas';
+import { rootReducers } from '../reducers';
+import { rootSaga } from '../sagas';
 
 const sagaMiddleware = createSagaMiddleware();
